Cascading Themes

Strong graphics packages are built on good defaults that are highly customizable. The CascadingTheme class makes it possible to define a list of defaults that specify inheritance, allowing customizable defaults with less repeated code.

Methods

Public methods

CascadingTheme$new()

CascadingTheme$compile()

CascadingTheme$keys()

CascadingTheme$inheritance()

CascadingTheme$value_validator()

CascadingTheme$node_validator()

CascadingTheme$node()

CascadingTheme$value()

CascadingTheme$value_base()

CascadingTheme$set_node()

CascadingTheme$set_value()

CascadingTheme$set_nodes()

CascadingTheme$set_values()

CascadingTheme$set_value_validator()

CascadingTheme$set_node_validator()

CascadingTheme$set_tree()

CascadingTheme$set_data()

CascadingTheme$set_node_validators()

CascadingTheme$set_value_validators()

CascadingTheme$clone()

Method new()

Usage

CascadingTheme$new(data = list())

Method compile()

Usage

CascadingTheme$compile()

Method keys()

Usage

CascadingTheme$keys()

Method inheritance()

Usage

CascadingTheme$inheritance(key)

Method value_validator()

Usage

CascadingTheme$value_validator(
  key,
  default = abort(glue::glue("No validators for value: '{key}'"))
)

Method node_validator()

Usage

CascadingTheme$node_validator(
  key,
  default = abort(glue::glue("No validators for node: '{key}'"))
)

Method node()

Usage

CascadingTheme$node(key, default = abort(glue::glue("No such node: '{key}'")))

Method value()

Usage

CascadingTheme$value(
  key,
  default = abort(glue::glue("No such value: '{key}'"))
)

Method value_base()

Usage

CascadingTheme$value_base(
  key,
  default = abort(glue::glue("No such node: '{key}'"))
)

Method set_node()

Usage

CascadingTheme$set_node(key, value)

Method set_value()

Usage

CascadingTheme$set_value(key, value)

Method set_nodes()

Usage

CascadingTheme$set_nodes(...)

Method set_values()

Usage

CascadingTheme$set_values(...)

Method set_value_validator()

Usage

CascadingTheme$set_value_validator(key, validator)

Method set_node_validator()

Usage

CascadingTheme$set_node_validator(key, validator)

Method set_tree()

Usage

CascadingTheme$set_tree(...)

Method set_data()

Usage

CascadingTheme$set_data(...)

Method set_node_validators()

Usage

CascadingTheme$set_node_validators(...)

Method set_value_validators()

Usage

CascadingTheme$set_value_validators(...)

Method clone()

The objects of this class are cloneable with this method.

Usage

CascadingTheme$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.