Coordinate systems

Coordinate systems pick which aesthetics are position aesthetics, finalize the Panel at the end of the build process, and transform coordinates. Regardless of their aesthetics, they should always return from their $transform() method a data frame with x and y columns. The default CoordIdentity does not change the Panel and performs no changes in its $transform() method.

Coord

CoordIdentity

Format

An object of class R6ClassGenerator of length 24.

Coord R6 Methods

Coord$new(aesthetics = c("x", "y"))

Create a coordinate system representing one or more position aesthetics.

coord$transform(data_mapped, scales)

Transform a data frame containing columns representing one or more position $aesthetics. This method should always return a data frame with x and y columns and leave non-position columns unchanged.

coord$finalize_panel(panel)

Finalize the Panel object at the end of the build process. This is where scale expansion (if any) can occur.

CoordIdentity inheritance

CoordIdentity inherits from Coord