Render and assemble a Graphic

The role of the renderer isn't solidified in ggr6. Currently it's an abstraction of various graphics backend that allow ggr6 to focus on the grammar (rather than on rendering graphical objects). Currently it has methods for rendering graphical primitives, methods for finding the aesthetics that are supported by each primitive, methods for constructing default scales for each aesthetic, and methods for assembling panels.

Details

Realistically, each renderer needs its own collection of Geom, Facet, Coord, and Scale subclasses. The role of ggr6 is to provide robust superclasses for these such that the subclasses can focus on rendering. This package provides RendererIdentity for testing and RendererGraphics, which uses the base graphics package to render graphics.

Methods

Public methods

Renderer$render_points()

Renderer$render_path()

Renderer$render_polygon()

Renderer$render_text()

Renderer$render_null()

Renderer$render_stack()

Renderer$render_panel()

Renderer$render_panels()

Renderer$default_scale()

Renderer$aesthetics()

Renderer$default_aesthetics()

Renderer$clone()

Method render_points()

Usage

Renderer$render_points(x, y, ...)

Method render_path()

Usage

Renderer$render_path(x, y, group, ...)

Method render_polygon()

Usage

Renderer$render_polygon(x, y, group, subgroup, ...)

Method render_text()

Usage

Renderer$render_text(x, y, label, ...)

Method render_null()

Usage

Renderer$render_null()

Method render_stack()

Usage

Renderer$render_stack(...)

Method render_panel()

Usage

Renderer$render_panel(panel, ...)

Method render_panels()

Usage

Renderer$render_panels(graphic, ...)

Method default_scale()

Usage

Renderer$default_scale(x, aesthetic)

Method aesthetics()

Usage

Renderer$aesthetics(geom_type)

Method default_aesthetics()

Usage

Renderer$default_aesthetics(geom_type)

Method clone()

The objects of this class are cloneable with this method.

Usage

Renderer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super class

ggr6::Renderer -> RendererIdentity

Methods

Public methods

RendererIdentity$render_points()

RendererIdentity$render_path()

RendererIdentity$render_polygon()

RendererIdentity$render_text()

RendererIdentity$render_null()

RendererIdentity$render_stack()

RendererIdentity$render_panel()

RendererIdentity$render_panels()

RendererIdentity$default_scale()

RendererIdentity$clone()

Inherited methods

ggr6::Renderer$aesthetics()

ggr6::Renderer$default_aesthetics()

Method render_points()

Usage

RendererIdentity$render_points(x, y, ...)

Method render_path()

Usage

RendererIdentity$render_path(x, y, group, ...)

Method render_polygon()

Usage

RendererIdentity$render_polygon(x, y, group, subgroup, ...)

Method render_text()

Usage

RendererIdentity$render_text(x, y, label, ...)

Method render_null()

Usage

RendererIdentity$render_null()

Method render_stack()

Usage

RendererIdentity$render_stack(...)

Method render_panel()

Usage

RendererIdentity$render_panel(panel, ...)

Method render_panels()

Usage

RendererIdentity$render_panels(graphic, ...)

Method default_scale()

Usage

RendererIdentity$default_scale(x, aesthetic)

Method clone()

The objects of this class are cloneable with this method.

Usage

RendererIdentity$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.