Graphics Renderer

A Renderer based on the graphics package. This is mostly used for testing at the moment.

Super class

ggr6::Renderer -> RendererGraphics

Methods

Public methods

RendererGraphics$render_points()

RendererGraphics$render_path()

RendererGraphics$render_polygon()

RendererGraphics$render_text()

RendererGraphics$render_null()

RendererGraphics$render_stack()

RendererGraphics$render_panel()

RendererGraphics$render_panels()

RendererGraphics$default_scale()

RendererGraphics$clone()

Inherited methods

ggr6::Renderer$aesthetics()

ggr6::Renderer$default_aesthetics()

Method render_points()

Usage

RendererGraphics$render_points(
  x,
  y,
  pch = 16,
  cex = 1,
  col = "black",
  lwd = 1,
  bg = "black",
  ...
)

Method render_path()

Usage

RendererGraphics$render_path(
  x,
  y,
  group = 1,
  lty = 1,
  lwd = 1,
  col = "black",
  ...
)

Method render_polygon()

Usage

RendererGraphics$render_polygon(
  x,
  y,
  group = 1,
  subgroup = 1,
  col = "black",
  fill = "grey50",
  lty = 1,
  ...
)

Method render_text()

Usage

RendererGraphics$render_text(
  x,
  y,
  label,
  cex = 1,
  col = "black",
  font = 1,
  ...
)

Method render_null()

Usage

RendererGraphics$render_null()

Method render_stack()

Usage

RendererGraphics$render_stack(...)

Method render_panel()

Usage

RendererGraphics$render_panel(panel, ...)

Method render_panels()

Usage

RendererGraphics$render_panels(graphic, ...)

Method default_scale()

Usage

RendererGraphics$default_scale(x, aesthetic)

Method clone()

The objects of this class are cloneable with this method.

Usage

RendererGraphics$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.