Convert between data spaces

Scales convert between the user data space, the data space used to compute the Stats (transformed data space), and the data space used by the Geoms (mapped data space) to draw graphical objects. Scales also compute breaks and labels used in the Guides. Most useful Scale classes inherit from ScaleSimple, which is designed to leverage the scales package. Each scale represents zero or more $aesthetics (character vector), and contains a $guide (Guide object) that will be used to communicate the relationship between user data values and mapped data values. ScaleNull is a scale that has infinite limits and never transforms or maps values, which is useful for testing and for values that should be passed to the Stat and Geom without being transformed or mapped.

Scale

ScaleNull

Format

An object of class R6ClassGenerator of length 24.

Scale R6 Methods

Scale$new(aesthetics = character(0))

Create a Scale object representing zero or more aesthetics.

scale$name(x)

Calculate the scale name, which should be a character vector of length 1. May return waiver() for when the scale name should be calculated from the ColumnMapping. This method may be removed in favour of this value living in the Guide.

scale$transform(x)

Transform the user-data vector x. Usually called by $transform_tbl().

scale$untransform(x)

Untransform the transformed-data vector x.

scale$train(x)

Train the scale with the transformed-data vector x. The range of values observed by the scale are used to calculate its $limits() (if these have not been specified manually). Usually called by $train_tbl().

scale$reset()

Forget all values that were 'remembered' by $train().

scale$map(x)

Convert the transformed-date vector x into mapped data space. This may be a numeric value (for position scales) or some other vector type (e.g., for colour scales). Usually called by $map_tbl().

scale$transform_tbl(data)

Call $transform() on the columns of data whose names are included in $aesthetics.

scale$untransform_tbl(data_trans)

Call $untransform() on the columns of data whose names are included in $aesthetics.

scale$train_tbl(data_trans)

Call $train() on the columns of data whose names are included in $aesthetics.

scale$map_tbl(data_trans)

Call $map() on the columns of data whose names are included in $aesthetics.

scale$breaks()

Calculate the breaks, which are usually nicely-numbered values around the scale $limits().

scale$breaks_minor()

Calculate the minor breaks, which are usually nicely-numbered values between the scale $breaks() and around the scale $limits().

scale$labels()

Calculate the labels associated with the $breaks().

scale$limits()

Calculate the scale $limits(), which is usually the range of values observed but could also be set.

scale$within_limits(x)

Returns a logical vector describing whether or not values in the transformed-data vector x are within the scale $limits().

scale$set_guide(guide)

Set the Guide that will be used to communicate the relationship between user data values and mapped data values.

ScaleNull inheritance

ScaleNull inherits from Scale