Discrete Scale

Discrete Scale

ScaleDiscrete

Format

An object of class R6ClassGenerator of length 24.

ScaleDiscrete inheritance

ScaleDiscrete inherits from ScaleSimple, Scale

ScaleDiscrete R6 Methods

scale_discrete$is_empty() (inherited from ScaleSimple)

Returns TRUE if the scale has $limits() (or can calculate them).

scale_discrete$breaks() (inherited from ScaleSimple)

Calculates and returns breaks in transformed data space. This can be a function of the $limits() and by default the $breaks() method of the transform is used. For continuous scales this is usually labeling::extended(), and for discrete scales the default is to return the limits unchanged.

scale_discrete$breaks_minor() (inherited from ScaleSimple)

Calculates and returns minor breaks in transformed data space. This can be a function of the $breaks(), $limits(), and number of intervals between breaks (usually 2). Minor breaks usually don't make sense for discrete scales.

scale_discrete$labels() (inherited from ScaleSimple)

Calculates and returns the labels associated with the $breaks(). This can be a function of the $breaks(), and by default the $format() method of the transform is used.

scale_discrete$limits() (inherited from ScaleSimple)

Calculates and returns the limits. This can be a funnction of the $trained_range(), and defaults to returning the $trained_range(). Continuous limits are always a numeric vector of length 2; discrete limits are usually a character vector containing possible values.

scale_discrete$trained_range() (inherited from ScaleSimple)

The range of the values in transformed data space that were observed for this scale's aesthetics. Contiuous ranges are always NULL (when no values were observed) or a numeric vector of length 2; discrete ranges are usually a character vector of values that were observed in the data.

scale_discrete$set_breaks(breaks) (inherited from ScaleSimple)

Set the breaks for this scale. Can be a vector of breaks in user data space or a function of the $limits().

scale_discrete$set_breaks_minor(breaks_minor) (inherited from ScaleSimple)

Set the minor breaks for this scale. Can be a vector of breaks in user data space or a function of the $breaks(), $limits(), and the number 2.

scale_discrete$set_labels(labels) (inherited from ScaleSimple)

Set the labels for this scale. Can be a vector of labels or a function of the $breaks().

scale_discrete$set_limits(limits) (inherited from ScaleSimple)

Set the limits for this scale. Can be a function of the $trained_range() or a vector (length 2 numeric for continuous scales or a character vector for discrete scales).

scale_discrete$set_oob(oob) (inherited from ScaleSimple)

Set the out-of-bounds function for this scale. The out-of-bounds function is a function of a vector and the $limits(); useful oob functions include oob_keep() (which does nothing; the default), scales::censor() (which sets values outside the limits to NA) and scales::squish() (which sets values outside the limits to the outermost values), and censor_discrete() (which works like scales::censor() but works for discrete scales).

scale_discrete$set_range(range) (inherited from ScaleSimple)

Set the range class for this scale. This must be a subclass of scales::Range.

scale_discrete$set_rescaler(rescaler) (inherited from ScaleSimple)

Set the rescaler function for this scale. Rescaled values (between 0 and 1) are often useful as inputs to continuous palletes. This function should take a keyword argument from, which is the scale $limits(). The most useful values for rescaler are rescale_none() (the default) and scales::rescale().

scale_discrete$set_trans(trans) (inherited from ScaleSimple)

Set the transform for this scale. See scales::trans_new() for a list of transforms for continuous scales. For discrete scales, the default value of discrete_identity_trans() is probably the only value that makes sense.

scale_discrete$set_na_value(na_value) (inherited from ScaleSimple)

Set the output (mapped data space) value that will represent NA values for this scale.

scale_discrete$set_limits_empty(limits_empty) (inherited from ScaleSimple)

Set the limits to return (in transformed data space) when $is_empty() returns TRUE.

scale_discrete$name(x) (inherited from Scale)

Calculate the scale name, which should be a character vector of length 1. May return waiver() for when the scale name should be calculated from the ColumnMapping. This method may be removed in favour of this value living in the Guide.

scale_discrete$map(x) (inherited from Scale)

Convert the transformed-date vector x into mapped data space. This may be a numeric value (for position scales) or some other vector type (e.g., for colour scales). Usually called by $map_tbl().

scale_discrete$within_limits(x) (inherited from Scale)

Returns a logical vector describing whether or not values in the transformed-data vector x are within the scale $limits().

scale_discrete$set_guide(guide) (inherited from Scale)

Set the Guide that will be used to communicate the relationship between user data values and mapped data values.