Mutable List of Scale objects

ScaleList objects contain a list of Scales, and have methods that can perform the same operation on all the scales they contain. They currently also generate the GuideList of trained Guides based on the Scales they contain.

Super class

ggr6::List -> ScaleList

Methods

Public methods

ScaleList$set()

ScaleList$aesthetics()

ScaleList$scale()

ScaleList$guides()

ScaleList$filter_by_aesthetics()

ScaleList$discard_by_aesthetics()

ScaleList$transform_tbl()

ScaleList$untransform_tbl()

ScaleList$train_tbl()

ScaleList$reset()

ScaleList$map_tbl()

ScaleList$add_missing()

ScaleList$clone()

Inherited methods

ggr6::List$add()

ggr6::List$get()

ggr6::List$initialize()

ggr6::List$size()

Method set()

Usage

ScaleList$set(index, item)

Method aesthetics()

Usage

ScaleList$aesthetics()

Method scale()

Usage

ScaleList$scale(aesthetic, default = NULL)

Method guides()

Usage

ScaleList$guides()

Method filter_by_aesthetics()

Usage

ScaleList$filter_by_aesthetics(aesthetics)

Method discard_by_aesthetics()

Usage

ScaleList$discard_by_aesthetics(aesthetics)

Method transform_tbl()

Usage

ScaleList$transform_tbl(data)

Method untransform_tbl()

Usage

ScaleList$untransform_tbl(data)

Method train_tbl()

Usage

ScaleList$train_tbl(data_trans)

Method reset()

Usage

ScaleList$reset()

Method map_tbl()

Usage

ScaleList$map_tbl(data_trans)

Method add_missing()

Usage

ScaleList$add_missing(data, renderer)

Method clone()

The objects of this class are cloneable with this method.

Usage

ScaleList$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.