Default implementation of Scale objects

Default implementation of Scale objects

ScaleSimple

Format

An object of class R6ClassGenerator of length 24.

ScaleSimple inheritance

ScaleSimple inherits from Scale

ScaleSimple R6 Methods

ScaleSimple$new(aesthetics = character(0))

Create a Scale object representing zero or more aesthetics.

scale_simple$name(x)

Calculate the scale name, which should be a character vector of length 1. May return waiver() for when the scale name should be calculated from the ColumnMapping. This method may be removed in favour of this value living in the Guide.

scale_simple$transform(x)

Transform the user-data vector x. Usually called by $transform_tbl().

scale_simple$untransform(x)

Untransform the transformed-data vector x.

scale_simple$train(x)

Train the scale with the transformed-data vector x. The range of values observed by the scale are used to calculate its $limits() (if these have not been specified manually). Usually called by $train_tbl().

scale_simple$reset()

Forget all values that were 'remembered' by $train().

scale_simple$map(x)

Convert the transformed-date vector x into mapped data space. This may be a numeric value (for position scales) or some other vector type (e.g., for colour scales). Usually called by $map_tbl().

scale_simple$transform_tbl(data)

Call $transform() on the columns of data whose names are included in $aesthetics.

scale_simple$untransform_tbl(data_trans)

Call $untransform() on the columns of data whose names are included in $aesthetics.

scale_simple$train_tbl(data_trans)

Call $train() on the columns of data whose names are included in $aesthetics.

scale_simple$map_tbl(data_trans)

Call $map() on the columns of data whose names are included in $aesthetics.

scale_simple$breaks()

Calculate the breaks, which are usually nicely-numbered values around the scale $limits().

scale_simple$breaks_minor() (inherited from Scale)

Calculate the minor breaks, which are usually nicely-numbered values between the scale $breaks() and around the scale $limits().

scale_simple$labels() (inherited from Scale)

Calculate the labels associated with the $breaks().

scale_simple$limits()

Calculate the scale $limits(), which is usually the range of values observed but could also be set.

scale_simple$within_limits(x)

Returns a logical vector describing whether or not values in the transformed-data vector x are within the scale $limits().

scale_simple$set_guide(guide) (inherited from Scale)

Set the Guide that will be used to communicate the relationship between user data values and mapped data values.