Transform data

Transform data

Transform data

Super class

ggr6::StatGeomBase -> Stat

Methods

Public methods

Stat$new()

Stat$set_default_mapping()

Stat$compute_panel()

Stat$compute_group()

Stat$clone()

Inherited methods

ggr6::StatGeomBase$aesthetics()

ggr6::StatGeomBase$apply_constant_aesthetic_values()

ggr6::StatGeomBase$apply_default_aesthetic_values()

ggr6::StatGeomBase$default_aesthetic_values()

ggr6::StatGeomBase$set_constant_aesthetic_values()

ggr6::StatGeomBase$setup_data()

Method new()

Usage

Stat$new(constant_aesthetic_values = list())

Method set_default_mapping()

Usage

Stat$set_default_mapping(default_mapping)

Method compute_panel()

Usage

Stat$compute_panel(data_trans, panel, renderer)

Method compute_group()

Usage

Stat$compute_group(data_trans, panel, renderer)

Method clone()

The objects of this class are cloneable with this method.

Usage

Stat$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

ggr6::StatGeomBase -> ggr6::Stat -> StatIdentity

Methods

Public methods

StatIdentity$compute_panel()

StatIdentity$clone()

Inherited methods

ggr6::StatGeomBase$aesthetics()

ggr6::StatGeomBase$apply_constant_aesthetic_values()

ggr6::StatGeomBase$apply_default_aesthetic_values()

ggr6::StatGeomBase$default_aesthetic_values()

ggr6::StatGeomBase$set_constant_aesthetic_values()

ggr6::StatGeomBase$setup_data()

ggr6::Stat$compute_group()

ggr6::Stat$initialize()

ggr6::Stat$set_default_mapping()

Method compute_panel()

Usage

StatIdentity$compute_panel(data_trans, panel, renderer)

Method clone()

The objects of this class are cloneable with this method.

Usage

StatIdentity$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.