Coordinate systems

Coordinate systems pick which aesthetics are position aesthetics, finalize the Panel at the end of the build process, and transform coordinates. Regardless of their aesthetics, they should always return from their $transform() method a data frame with x and y columns. The default CoordIdentity does not change the Panel and performs no changes in its $transform() method.

Methods

Public methods

Coord$new()

Coord$transform()

Coord$finalize_panel()

Coord$clone()

Method new()

Usage

Coord$new(aesthetics = c("x", "y"))

Method transform()

Usage

Coord$transform(data_mapped, scales)

Method finalize_panel()

Usage

Coord$finalize_panel(panel)

Method clone()

The objects of this class are cloneable with this method.

Usage

Coord$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super class

ggr6::Coord -> CoordIdentity

Methods

Public methods

CoordIdentity$transform()

CoordIdentity$finalize_panel()

CoordIdentity$clone()

Inherited methods

ggr6::Coord$initialize()

Method transform()

Usage

CoordIdentity$transform(data_mapped, scales)

Method finalize_panel()

Usage

CoordIdentity$finalize_panel(panel)

Method clone()

The objects of this class are cloneable with this method.

Usage

CoordIdentity$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.