Graphical represntation of Scales

Guide objects are a graphical representation of Scales, in that they communicate the relationship between user data values and mapped data values. Each Scale must have a Guide, although this can be (and often is) GuideNull, which draws nothing). Guides which have the same name, class, breaks, and labels are merged together, such that in the final plot there may be more than one scale represented by one guide. All guides except GuideNull will probably have to be subclassed for each Renderer.

Guide

GuideNull

Format

An object of class R6ClassGenerator of length 24.

Guide R6 Methods

Guide$new()

Create a Guide object.

guide$title()

Return the title that should be displayed by this guide.

guide$position()

Return the object describing the position of this guide (specific to the Renderer). A value of NULL means it should not be displayed; a value of waiver() means the position is unspecified.

guide$aesthetics()

The aesthetics represented by this Guide.

guide$train(scale)

Add information from a Scale to this guide. This creates the $key field, which is a tibble::tibble() with a column for each aesthetic and a .breaks and .labels column.

guide$train_layers(layers, renderer)

Add information from a LayerList to this object. This is used to assemble the default aesthetic values and geometry primatives that will be displayed by the guide.

guide$merge(guide)

Merges information from another guide into this guide. The default implementation is to merge only if the Guides have the same class, breaks, labels, title, and position. This would occur if a user maps the same column to two aesthetics. Returns TRUE if a merge occured, FALSE otherwise. This is usually called by GuideList's $merge_all() method.

guide$render(panel, renderer)

Renders the guide.

guide$set_key(key)

Sets the $key for this Guide, which is a tibble::tibble() with a column for each aesthetic and a .breaks and .labels column.

guide$set_position(position)

Sets the $position() for this guide. Use NULL to hide, waiver() to let the Renderer choose, or some other Renderer-specific value.

guide$make_key(scale, censor = TRUE)

Creates the $key based on a Scale, optionally censoring the breaks to ensure they are within the Scale's $limits().

GuideNull inheritance

GuideNull inherits from Guide